Add null check for product image URL before deletion
This update introduces a check for null or whitespace in
`existProduct.ImgUrl` prior to file deletion. This prevents
errors when attempting to delete a non-existent image file,
ensuring that deletion only occurs with a valid image URL.

+++ Areas/Admin/Services/ProductService/ProductService.cs
@@ -139,7 +139,8 @@ public class ProductService : IProductService
 			// nếu có chỉnh sửa ảnh đại diện của sản phẩm
 			if (request.AvatarUpdate is not null)
 			{
-				File.Delete(@"wwwroot\img\products\" + existProduct.ImgUrl);
+				if (!string.IsNullOrWhiteSpace(existProduct.ImgUrl))
+					File.Delete(@"wwwroot\img\products\" + existProduct.ImgUrl);
 
 				existProduct.ImgUrl = SetFileName(request.AvatarUpdate);
 				await UploadImageAssync(request.AvatarUpdate);
